fix: 미루기 전 미루기 횟수 주는 API
