+++ src/controller/plantController.js
@@ -111,39 +111,6 @@ module.exports = {
     }
   },
 
-  getWaterPossible: async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        errors: errors.array(),
-      });
-    }
-    const { CherishId } = req.query;
-    const t = await sequelize.transaction();
-    try {
-      const cherish = await Cherish.findOne({
-        attributes: ['postpone_number'],
-        where: {
-          id: CherishId,
-        },
-      });
-      const is_limit_postpone_number = cherish && cherish.postpone_number >= 3 ? true : false;
-
-      // 이미 미루기 횟수가 초과된 경우
-      if (is_limit_postpone_number) {
-        return res.status(sc.OK).send(ut.fail(rm.IMPOSSIBLE_WATER_POSTPONE));
-      }
-      return res.status(sc.OK).send(
-        ut.success(rm.POSSIBLE_WATER_POSTPONE, {
-          cherish,
-        })
-      );
-    } catch (err) {
-      console.log(err);
-      return res.status(sc.INTERNAL_SERVER_ERROR).send(ut.fail(rm.INTERNAL_SERVER_ERROR));
-    }
-  },
-
   getCherishInfo: async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
