+++ src/modules/responseMessage.js
@@ -8,7 +8,7 @@ module.exports = {
   CANT_ACCESS: '접근할 수 없습니다.',
 
   IMPOSSIBLE_WATER_POSTPONE: '이미 미루기 횟수를 초과하셨습니다.',
-  POSSIBLE_WATER_POSTPONE: '미루기가 가능합니다.',
+  GET_WATER_POSTPONE: '미루기 횟수입니다.',
   READ_ALL_CHERISH_BY_ID_SUCCESS: '체리쉬 상세보기를 성공하였습니다.',
 
   /* 로그인 */
